requirements.txt
dgl==0.3
scipy==1.1.0
torch_scatter==2.0.6
torch-cluster==1.5.9
torch-sparse==0.6.9
torch_geometric==1.4.3
numpy==1.19.5
hyperopt==0.2.5
scikit_learn==0.21.3
requests